Move format_work_data to where its used
